requirements.txt
pip
autopep8
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
requests
pyngrok
pycryptodome
bs4
google-generativeai
langchain-google-genai
langchain
pydantic